ColumnSearchPropertyField moved to shared-library
