--- SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/ColumnFormPanel/ColumnSearchPropertyField/types.ts
@@ -1,5 +1,0 @@
-import { ITextFieldProps } from '@fluentui/react'
-
-export interface IColumnSearchPropertyFieldProps extends Omit<ITextFieldProps, 'onChange'> {
-  onChange: (value: string) => void
-}
